refactor: move types closer to search
